feat: refactor oidc validator
simplifies the oidc validator
