+++ bonsai-pay/zkvm/methods/Cargo.toml
@@ -9,7 +9,7 @@ methods = ["guest"]
 [dev-dependencies]
 alloy-primitives = "0.3.3"
 alloy-sol-types = "0.3.2"
-oidc_validate = { path = "../oidc_validate/" }
+oidc-validator = { path = "../oidc-validator/" }
 risc0-zkvm = { version = "0.19.0" }
 
 [build-dependencies]
