+++ bonsai-pay/zkvm/methods/guest/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 risc0-zkvm = { version = "0.19.0", features = ['std'], default-features = false }
-oidc_validate = { path = "../../oidc_validate/"}
+oidc-validator = { path = "../../oidc-validator/"}
 alloy-sol-types = "0.3.2"
 alloy-primitives = "0.3.3"
 hex = "0.4.3"
