+++ bonsai-pay/zkvm/methods/guest/src/main.rs
@@ -3,7 +3,7 @@ use std::str::FromStr;
 
 use alloy_primitives::Address;
 use alloy_sol_types::SolType;
-use oidc_validate::IdentityProvider;
+use oidc_validator::IdentityProvider;
 use risc0_zkvm::guest::env;
 use risc0_zkvm::sha::rust_crypto::{Digest as _, Sha256};
