+++ bonsai-pay/zkvm/oidc-validator/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "oidc_validate"
+name = "oidc-validator"
 version = "0.1.0"
 edition = "2021"
